+++ models/eppi/Document.go
@@ -9,13 +9,13 @@ import (
 )
 
 type Document struct {
-	Id           uint16   `json:"id"`
+	Id           int      `json:"id"`
 	Abstract     string   `json:"abstract"`
 	Additional   string   `json:"additional"`
 	Appointed    Date     `json:"appointed"`
-	AuthRef      uint16   `json:"authRef"`
+	AuthRef      int      `json:"authRef"`
 	Breviate     Breviate `json:"breviate"`
-	BreviatePage uint16   `json:"breviatePage"`
+	BreviatePage int      `json:"breviatePage"`
 	Chairman     Person   `json:"chairman"`
 	CorpAuthors  string   `json:"corpAuthors"`
 
@@ -23,7 +23,7 @@ type Document struct {
 	Doctype   string      `json:"doctype"`
 	HeldBy    Institution `json:"heldBy"`
 	LCSubject Subject     `json:"lcSubject"`
-	Sectioned uint16      `json:"sectioned"`
+	Sectioned int         `json:"sectioned"`
 
 	// Paper number is arabic, but some documents contain
 	// multiple paper numbers.
@@ -35,7 +35,7 @@ type Document struct {
 	// Session might be better included in breviates instead.
 	Session   Session `json:"session"`
 	Source    string  `json:"source"`
-	StartPage uint16  `json:"startPage"`
+	StartPage int     `json:"startPage"`
 	Subvol    string  `json:"Subvol"`
 	Tables    string  `json:"tables"`
 	Title     string  `json:"title"`
